build: make workspace builds reproducible across toolchain versions
Pins the exact compiler in rust-toolchain.toml (1.98.0, no floating
channel), adds scripts/check-toolchain.sh to enforce the reproducible-
build policy (exact pin, required wasm target/components, CI consuming
the pin, committed Cargo.lock files, active rustc matching the pin),
and makes every build/test entry point resolve dependencies --locked
against committed lock files. CI now derives its toolchain from
rust-toolchain.toml instead of a floating stable ref, runs clippy with
-D warnings, and fails fast on policy drift before compiling.

Also fixes the lint failures surfaced by the new clippy gate
(oracles.rs doc-comment artifact, op_ref; unused imports and
unnecessary casts in tests) and removes the stray duplicate braces
left in contracts/credit/src/types.rs by a prior merge (one of the
documented merge-artifact errors; the remaining legacy workspace
errors are tracked separately and are untouched).

Closes #1163

+++ contracts/creditra-credit/tests/snap_prorate.rs
@@ -640,7 +640,7 @@ mod fuzz {
 
     /// Elapsed seconds, 0 to 100 years.
     fn elapsed_secs() -> impl Strategy<Value = u64> {
-        0u64..=(SECONDS_PER_YEAR as u64 * 100)
+        0u64..=(SECONDS_PER_YEAR * 100)
     }
 
     proptest! {
@@ -764,7 +764,7 @@ mod fuzz {
         fn interest_bounded_by_principal_within_one_year(
             p in safe_principal(),
             r in rate_bps(),
-            t in 0u64..=(SECONDS_PER_YEAR as u64),
+            t in 0u64..=SECONDS_PER_YEAR,
         ) {
             if let Ok(interest) = accrued_interest(Uint128::new(p), r, t) {
                 prop_assert!(
@@ -781,8 +781,8 @@ mod fuzz {
         fn split_period_accrues_le_combined(
             p in safe_principal(),
             r in rate_bps(),
-            t1 in 0u64..=(SECONDS_PER_YEAR as u64),
-            t2 in 0u64..=(SECONDS_PER_YEAR as u64),
+            t1 in 0u64..=SECONDS_PER_YEAR,
+            t2 in 0u64..=SECONDS_PER_YEAR,
         ) {
             let principal = Uint128::new(p);
             let combined = t1.saturating_add(t2);
